+++ crates/deadpool/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add unique `id` to objects which can be read via the `Object::id` method
+
 ## [0.12.2] - 2025-02-02
 
 - Update `itertools` dependency to version `0.13.0`